+++ packages/node-stdlib/src/bytes/mod.ts
@@ -2,3 +2,4 @@
 // All rights reserved. MIT License.
 
 export * from "./bytes";
+export * from "./dynamic_buffer";
