+++ packages/node-stdlib/test/bytes/dynamic_buffer.test.ts
@@ -1,10 +1,10 @@
 import { inspect } from "util";
-import { core, util } from "../../src";
+import { bytes, util } from "../../src";
 
-describe("core/dynamic_buffer.ts", () => {
+describe("bytes/dynamic_buffer.ts", () => {
   describe("constructor()", () => {
     it("creates an empty buffer", () => {
-      const buf = new core.DynamicBuffer();
+      const buf = new bytes.DynamicBuffer();
       expect(buf.isEmpty).toBe(true);
       expect(buf.length).toBe(0);
       expect(buf.capacity).toBe(64);
@@ -14,7 +14,7 @@ describe("core/dynamic_buffer.ts", () => {
 
     it("creates a DynamicBuffer from the given Buffer", () => {
       const b = Buffer.from("abc");
-      const buf = new core.DynamicBuffer(b);
+      const buf = new bytes.DynamicBuffer(b);
       expect(buf.isEmpty).toBe(false);
       expect(buf.length).toBe(3);
       expect(buf.capacity).toBe(3);
@@ -23,7 +23,7 @@ describe("core/dynamic_buffer.ts", () => {
     });
 
     it("creates a DynamicBuffer from the given string", () => {
-      const buf = new core.DynamicBuffer("10ff", "hex");
+      const buf = new bytes.DynamicBuffer("10ff", "hex");
       expect(buf.isEmpty).toBe(false);
       expect(buf.length).toBe(2);
       expect(buf.capacity).toBe(2);
@@ -36,7 +36,7 @@ describe("core/dynamic_buffer.ts", () => {
     it("resets the buffer when n is 0", () => {
       const b = Buffer.from(new ArrayBuffer(8));
       b.write("10ffab23ef5d", "hex");
-      const buf = new core.DynamicBuffer(b.buffer);
+      const buf = new bytes.DynamicBuffer(b.buffer);
       expect(buf.length).toBe(8);
       expect(buf.capacity).toBe(8);
       buf.truncate(0);
@@ -47,7 +47,7 @@ describe("core/dynamic_buffer.ts", () => {
     it("truncates all but the first byte", () => {
       const b = Buffer.from(new ArrayBuffer(8));
       b.write("10ffab23ef5d", "hex");
-      const buf = new core.DynamicBuffer(b.buffer);
+      const buf = new bytes.DynamicBuffer(b.buffer);
       expect(buf.length).toBe(8);
       expect(buf.capacity).toBe(8);
       buf.truncate(3);
@@ -59,7 +59,7 @@ describe("core/dynamic_buffer.ts", () => {
     it("panics if n is out of range", () => {
       const b = Buffer.from(new ArrayBuffer(8));
       b.write("10ffab23ef5d", "hex");
-      const buf = new core.DynamicBuffer(b.buffer);
+      const buf = new bytes.DynamicBuffer(b.buffer);
       expect(() => {
         buf.truncate(9);
       }).toPanic("DynamicBuffer: truncation out of range");
@@ -70,7 +70,7 @@ describe("core/dynamic_buffer.ts", () => {
     it("resets the buffer", () => {
       const b = Buffer.from(new ArrayBuffer(8));
       b.write("10ffab23ef5d", "hex");
-      const buf = new core.DynamicBuffer(b.buffer);
+      const buf = new bytes.DynamicBuffer(b.buffer);
       expect(buf.length).toBe(8);
       expect(buf.capacity).toBe(8);
       buf.reset();
@@ -83,7 +83,7 @@ describe("core/dynamic_buffer.ts", () => {
     it("grows the buffer to the specified capacity", () => {
       const b = Buffer.from(new ArrayBuffer(8));
       b.write("10ffab23ef5d", "hex");
-      const buf = new core.DynamicBuffer(b.buffer);
+      const buf = new bytes.DynamicBuffer(b.buffer);
       expect(buf.length).toBe(8);
       expect(buf.capacity).toBe(8);
       buf.grow(8);
@@ -94,7 +94,7 @@ describe("core/dynamic_buffer.ts", () => {
     it("panics if n is negative", () => {
       const b = Buffer.from(new ArrayBuffer(8));
       b.write("10ffab23ef5d", "hex");
-      const buf = new core.DynamicBuffer(b.buffer);
+      const buf = new bytes.DynamicBuffer(b.buffer);
       expect(() => {
         buf.grow(-1);
       }).toPanic("DynamicBuffer.grow: negative count");
@@ -103,34 +103,34 @@ describe("core/dynamic_buffer.ts", () => {
 
   describe("write()/writeString()/writeByte()", () => {
     it("writes the buffer contents to the DynamicBuffer", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       buf.write(Buffer.from("10ffab", "hex"));
       expect(buf.length).toBe(9);
       expect(buf.toString("hex")).toBe("10ffab23ef5d10ffab");
     });
 
     it("writes the string to the DynamicBuffer", () => {
-      const buf = new core.DynamicBuffer("Hello");
+      const buf = new bytes.DynamicBuffer("Hello");
       buf.writeString(" world!");
       expect(buf.toString()).toBe("Hello world!");
     });
 
     it("writes the hex string to the DynamicBuffer", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       buf.writeString("10ffab", "hex");
       expect(buf.length).toBe(9);
       expect(buf.toString("hex")).toBe("10ffab23ef5d10ffab");
     });
 
     it("writes the byte to the DynamicBuffer", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       buf.writeByte(0x10);
       expect(buf.length).toBe(7);
       expect(buf.toString("hex")).toBe("10ffab23ef5d10");
     });
 
     it("panics if c isn't a valid byte", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       expect(() => {
         buf.writeByte(1.1);
       }).toPanic("DynamicBuffer.writeByte: c is not a valid byte");
@@ -139,7 +139,7 @@ describe("core/dynamic_buffer.ts", () => {
 
   describe("read()", () => {
     it("reads bytes and puts them in the buffer", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const b = Buffer.alloc(4);
       const r = buf.read(b);
       expect(r.isSuccess());
@@ -149,7 +149,7 @@ describe("core/dynamic_buffer.ts", () => {
     });
 
     it("reads all the bytes if the buffer is larger than the DynamicBuffer", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const b = Buffer.alloc(10);
       const r = buf.read(b);
       expect(r.isSuccess());
@@ -159,17 +159,17 @@ describe("core/dynamic_buffer.ts", () => {
     });
 
     it("returns eof if the buffer is empty", () => {
-      const buf = new core.DynamicBuffer();
+      const buf = new bytes.DynamicBuffer();
       const b = Buffer.alloc(10);
       const r = buf.read(b);
       expect(r.isFailure());
-      expect(r.failure()).toBe(core.eof);
+      expect(r.failure()).toBe(bytes.eof);
     });
   });
 
   describe("next()", () => {
     it("returns a buffer with the next n bytes", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const b = buf.next(4);
       expect(b.length).toBe(4);
       expect(b.toString("hex")).toBe("10ffab23");
@@ -177,7 +177,7 @@ describe("core/dynamic_buffer.ts", () => {
     });
 
     it("returns all the bytes when n is greater then the buffer length", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const b = buf.next(10);
       expect(b.length).toBe(6);
       expect(b.toString("hex")).toBe("10ffab23ef5d");
@@ -187,7 +187,7 @@ describe("core/dynamic_buffer.ts", () => {
 
   describe("readByte()", () => {
     it("reads the next byte", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const r = buf.readByte();
       expect(r.isSuccess());
       expect(r.success()).toBe(0x10);
@@ -195,16 +195,16 @@ describe("core/dynamic_buffer.ts", () => {
     });
 
     it("returns eof if the buffer is empty", () => {
-      const buf = new core.DynamicBuffer();
+      const buf = new bytes.DynamicBuffer();
       const r = buf.readByte();
       expect(r.isFailure());
-      expect(r.failure()).toBe(core.eof);
+      expect(r.failure()).toBe(bytes.eof);
     });
   });
 
   describe("readBytes()", () => {
     it("reads bytes until the delim", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const [b, err] = buf.readBytes(0xab);
       expect(err).toBeUndefined();
       expect(b.toString("hex")).toBe("10ffab");
@@ -212,15 +212,15 @@ describe("core/dynamic_buffer.ts", () => {
     });
 
     it("reads the whole buffer and returns eof if delim doesn't exist", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const [b, err] = buf.readBytes(0xcc);
-      expect(err).toBe(core.eof);
+      expect(err).toBe(bytes.eof);
       expect(b.toString("hex")).toBe("10ffab23ef5d");
       expect(buf.length).toBe(0);
     });
 
     it("panics if delim is not a valid byte", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       expect(() => {
         buf.readBytes(1.1);
       }).toPanic("DynamicBuffer.readBytes: delim is not a valid byte");
@@ -229,7 +229,7 @@ describe("core/dynamic_buffer.ts", () => {
 
   describe("readString()", () => {
     it("reads bytes until the delim as a string", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const [s, err] = buf.readString(0xab, "hex");
       expect(err).toBeUndefined();
       expect(s).toBe("10ffab");
@@ -237,7 +237,7 @@ describe("core/dynamic_buffer.ts", () => {
     });
 
     it("panics if delim is not a valid byte", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       expect(() => {
         buf.readString(1.1);
       }).toPanic("DynamicBuffer.readBytes: delim is not a valid byte");
@@ -246,7 +246,7 @@ describe("core/dynamic_buffer.ts", () => {
 
   describe("@@iterator()", () => {
     it("reads the buffer by using for of", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const arr: number[] = [];
       for (const b of buf) {
         arr.push(b);
@@ -259,7 +259,7 @@ describe("core/dynamic_buffer.ts", () => {
 
   describe("Copyable conformance", () => {
     it("creates a copy of the DynamicBuffer", () => {
-      const buf = new core.DynamicBuffer("abc");
+      const buf = new bytes.DynamicBuffer("abc");
       const bufCopy = util.copy(buf);
       bufCopy.writeString("def");
       expect(buf).not.toBe(bufCopy);
@@ -270,13 +270,13 @@ describe("core/dynamic_buffer.ts", () => {
 
   describe("inspect", () => {
     it("returns just the type when depth is zero", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const s = inspect(buf, { depth: -1 });
       expect(s).toBe("DynamicBuffer {}");
     });
 
     it("returns a string representation of the box", () => {
-      const buf = new core.DynamicBuffer("10ffab23ef5d", "hex");
+      const buf = new bytes.DynamicBuffer("10ffab23ef5d", "hex");
       const s = inspect(buf);
       expect(s).toBe("DynamicBuffer { 10 ff ab 23 ef 5d }");
     });
